refactor: rename to logSkillHint.ts and move the trigger logic there
Review feedback on #3501: the module and function are now named logSkillHint,
and the timers/middleware that schedule the check live in the module instead of
pluginDev.ts, whose configureServer hook is now a one-liner.
