+++ packages/vike/src/node/vite/plugins/pluginDev.ts
@@ -4,10 +4,9 @@ export { logDockerHint }
 import { type Plugin, type ResolvedConfig, type UserConfig } from 'vite'
 import { optimizeDeps, resolveOptimizeDeps } from './pluginDev/optimizeDeps.js'
 import { determineFsAllowList } from './pluginDev/determineFsAllowList.js'
-import { checkVikeSkill } from './pluginDev/vikeSkill.js'
+import { logSkillHint } from './pluginDev/logSkillHint.js'
 import { addSsrMiddleware } from '../shared/addSsrMiddleware.js'
 import { isDebugError } from '../../../utils/debug.js'
-import { setTimeoutUnref } from '../../../utils/setTimeoutUnref.js'
 import { applyDev } from '../../../utils/isDev.js'
 import { isDocker } from '../../../utils/isDocker.js'
 import { assertWarning } from '../../../utils/assert.js'
@@ -41,29 +40,7 @@ function pluginDev(): Plugin[] {
       },
       configureServer: {
         handler(server) {
-          // Check whether the user installed Vike's skill for AI agents (https://vike.dev/ai#install) — late, so that it never slows down dev start nor the first page requests: 5 seconds after the first request, or at most 10 seconds after the server started.
-          let isDone = false
-          const runAfter = (milliseconds: number) => {
-            setTimeoutUnref(() => {
-              if (isDone) return
-              isDone = true
-              checkVikeSkill(config.root)
-            }, milliseconds)
-          }
-          if (server.httpServer) {
-            server.httpServer.once('listening', () => runAfter(10 * 1000))
-          } else {
-            // Middleware mode: the HTTP server is owned by the user
-            runAfter(10 * 1000)
-          }
-          let isFirstRequest = true
-          server.middlewares.use((_req, _res, next) => {
-            if (isFirstRequest) {
-              isFirstRequest = false
-              runAfter(5 * 1000)
-            }
-            next()
-          })
+          logSkillHint(server, config.root)
         },
       },
     },
